fix: block and metadata update
